Introducing image attachments
Before returning a response, check if the 'attachments' property is
defined. If yes, load the file (for the moment only images) in the
required variation, get the link and build a Markdown representation.
Search in the 'field' specified in the attachment and replace the
'slug' tag with the generated link.