Introducing 00_utils.js
Utilities that can be used accross all Pillar apps

This includes general purpose functions,
small jQuery plugins, and so on.

For example: the autoResize jQuery plugin that automatically
resizes a textarea according to its content was used on all
major Pillar apps so far (Attract, Flamenco, Blender Cloud
and Dillo), proven to be really needed everywhere.